# app/src/generated/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/settings/tokens/{id}": {
      "delete": {
        "operationId": "TokensApi_removeToken",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "TokensApi"
        ]
      }
    },
    "/settings/tokens/{id}/permissions": {
      "post": {
        "operationId": "TokensApi_addPermission",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "TokensApi"
        ]
      }
    },
    "/settings/tokens/{id}/permissions/{index}": {
      "delete": {
        "operationId": "TokensApi_removePermission",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "index",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "TokensApi"
        ]
      }
    },
...
  },
  "info": {
    "title": "7router API",
    "description": "Unified cloud storage management API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}